+++ beta/module/install_nga.sh
@@ -137,38 +137,28 @@ if [ -d /data/data/com.google.android.googlequicksearchbox ] && [ $API -ge 29 ] 
         fi
 # Make Google app as system app
         if [ -z $(pm list packages -s com.google.android.googlequicksearchbox | grep -v nga) ] && [ ! -f /data/adb/modules/PixelifyNext/system/product/priv-app/Velvet/Velvet.apk ] || [ $FORCE_VELVET -eq 1 ]; then
-            error "- Google is not installed as a system app !!"
-            log "- Making Google a system app"
-            log " - Making Google a system app"
-            log ""
-            if [ -f $app/com.google.android.googlequicksearchbox*/base.apk ]; then
-                cp -r $app/com.google.android.googlequicksearchbox*/. $MODPATH/system/product/priv-app/Velvet
-                mv $MODPATH/system/product/priv-app/Velvet/base.apk $MODPATH/system/product/priv-app/Velvet/Velvet.apk
-            else
-                cp -r /data/adb/modules/PixelifyNext/system$product/priv-app/Velvet/. $MODPATH/system$product/priv-app/Velvet
-            fi
-            rm -rf $MODPATH/system/product/priv-app/Velvet/oat
-            #mv $MODPATH/files/privapp-permissions-com.google.android.googlequicksearchbox.xml $MODPATH/system/product/etc/permissions/privapp-permissions-com.google.android.googlequicksearchbox.xml
-
-        # If Pixelify made system app then remake it
-        elif [ -f /data/adb/modules/PixelifyNext/system/product/priv-app/Velvet/Velvet.apk ]; then
-            if [ $FORCE_VELVET -eq 2 ]; then
-                error "- Google is not installed as a system app !!"
-                log "- Making Google a system app"
+            VELVET_DIR="$(get_pkg_app_dir "com.google.android.googlequicksearchbox")"
+            if [ -n "$VELVET_DIR" -a -d "$VELVET_DIR" ]; then
+                ui_print "- Making Google a system app"
                 log " - Making Google a system app"
-                log ""
-                if [ -f $app/com.google.android.googlequicksearchbox*/base.apk ]; then
-                    cp -r $app/com.google.android.googlequicksearchbox*/. $MODPATH/system/product/priv-app/Velvet
-                    mv $MODPATH/system/product/priv-app/Velvet/base.apk $MODPATH/system/product/priv-app/Velvet/Velvet.apk
-                else
-                    cp -r data/adb/modules/Pixelify/system$product/priv-app/Velvet/. $MODPATH/system$product/priv-app/Velvet
-                fi
-                rm -rf $MODPATH/system/product/priv-app/Velvet/oat
+                mkdir -p $MODPATH/system/product/priv-app/Velvet 2>/dev/null
+                cp -r "$VELVET_DIR/." $MODPATH/system/product/priv-app/Velvet 2>/dev/null || true
+                [ -f $MODPATH/system/product/priv-app/Velvet/base.apk ] && mv $MODPATH/system/product/priv-app/Velvet/base.apk $MODPATH/system/product/priv-app/Velvet/Velvet.apk 2>/dev/null || true
+                rm -rf $MODPATH/system/product/priv-app/Velvet/oat 2>/dev/null || true
             fi
-            #mv $MODPATH/files/privapp-permissions-com.google.android.googlequicksearchbox.xml $MODPATH/system/product/etc/permissions/privapp-permissions-com.google.android.googlequicksearchbox.xml
+        elif [ -f /data/adb/modules/PixelifyNext/system/product/priv-app/Velvet/Velvet.apk ]; then
+            VELVET_DIR="$(get_pkg_app_dir "com.google.android.googlequicksearchbox")"
+            if [ -n "$VELVET_DIR" -a -d "$VELVET_DIR" ]; then
+                ui_print "- Making Google a system app"
+                log " - Making Google a system app"
+                mkdir -p $MODPATH/system/product/priv-app/Velvet 2>/dev/null
+                cp -r "$VELVET_DIR/." $MODPATH/system/product/priv-app/Velvet 2>/dev/null || true
+                [ -f $MODPATH/system/product/priv-app/Velvet/base.apk ] && mv $MODPATH/system/product/priv-app/Velvet/base.apk $MODPATH/system/product/priv-app/Velvet/Velvet.apk 2>/dev/null || true
+                rm -rf $MODPATH/system/product/priv-app/Velvet/oat 2>/dev/null || true
+            fi
         fi
     else
-     NO_NGA=true
+        NO_NGA=true
     fi
 fi
 }
